JUNIE Implement createIndexRequests function
Implement a kotlin function that accepts a meta-model and returns a list of CreateIndexRequest instances. There should be one instance for each entity in the meta-model. The index name should be a concatenation of the package name and the entity name (use __ as the separator). The function should use metaModelForEach() to implement the functionality (see other implementations that use this function to get an idea). The function should be in a package named org.treeWare.elasticsearch.index

Plan:
1. Explore the tree-ware-kotlin-elasticsearch project structure to understand the current codebase
2. Look for examples of metaModelForEach() usage in the codebase to understand its implementation pattern
3. Examine the Leader1MetaModelVisitor interface and examples of its implementation
4. Check the CreateIndexRequest class in the Elasticsearch Java client
5. Create the necessary directory structure for the org.treeWare.elasticsearch.index package
6. Implement the function to generate CreateIndexRequest instances for each entity in the meta-model
7. Write tests for the new function
8. Run the build to ensure there are no errors
9. Commit the changes with the required commit message format

Time: 1 hour